graph more accurate now
