--- src/ros/nova-gui/nova-gui/src/components/NIRProbe/NIRProbeCalibration/DefaultWaterCalibration.ts
@@ -1,20 +1,0 @@
-// TODO find a better solution for this
-export const DEFAULT_WATER_CALIBRATION = {
-  "points": [
-    {"difference":3029.5,"concentration":0},
-    {"difference":2245.666667,"concentration":2.5},
-    {"difference":2663.666667,"concentration":5},
-    {"difference":2311.833333,"concentration":7.5},
-    {"difference":1638,"concentration":10},
-    {"difference":1262.333333,"concentration":12.5},
-    {"difference":1494.5,"concentration":15},
-    {"difference":1247,"concentration":17.5},
-    {"difference":1304.666667,"concentration":20},
-    {"difference":886.5,"concentration":22.5},
-    {"difference":987,"concentration":25},
-    {"difference":571.3333333,"concentration":30}
-  ],
-  "yIntercept":0.00875,
-  "gradient":0.0218,
-  "chemBlankDifference":2155.833333
-};
