+++ src/ros/nova-gui/nova-gui/src/components/NIRProbe/NIRProbeCalibration/NIRCalibration.ts
@@ -15,10 +15,9 @@ export const calibrationFunction = (c: number[]) => (x: number, y: number): numb
     return 0
   }
 
-  return c[0] * Math.log10(x)
-    + c[1] * Math.log10(y)
-    + c[2] * Math.log10(x + y)
-    + c[3] * Math.log10(x - y)
-    + c[4] * Math.exp(x)
-    + c[5]
+  return (c[0] * Math.log10(x+1)
+    + c[1] * Math.log10(y+1)
+    // + c[2] * Math.log10(x + y)
+    // + c[3] * Math.log10(x - y)
+    + c[5]) % 100
 }
