+++ src/ros/nova-gui/nova-gui/src/components/NIRProbe/NIRProbeOutputWidget/NIRProbeWidget.tsx
@@ -5,7 +5,7 @@ import NIRProbeFileTableWidget from "./NIRProbeFileTableWidget.tsx";
 import NIRCalibrationCurveWidget, {EMPTY_CALIBRATION_DATA, NIRCalibrationData} from "../NIRProbeCalibration/NIRCalibrationCurveWidget.tsx";
 import {useLocalStorage} from "../../../hooks/useLocalStorage.ts";
 import TOFHeight from "../../AnalysisPlatformHeight/AnalysisPlatformHeight.tsx";
-import {DEFAULT_WATER_CALIBRATION} from "../NIRProbeCalibration/DefaultWaterCalibration.ts";
+import {DEFAULT_WATER_CALIBRATION} from "../NIRProbeCalibration/DefaultCalibrationPoints.ts";
 import {useGenericStore} from "../../../hooks/useGenericStore.ts";
 import {Site} from "../../../redux/models/genericStores/CurrentSiteStore.ts";
 import {SiteData, SiteDataState} from "../../../redux/models/genericStores/SiteDataState.ts";
